New sensors added / enabled.
Used the heatingMode sensor that was commented out and changed it to a TextSensor. This displays what you would expect to see on the iBoost including a message for when the tank is hot thanks to the cylinderHot  variable that was also commented out.
Added a new sensor heatingImport based on the W1 value from the debug. This matches the value from the CT clamp and is positive when importing power and negative when exporting.
Added new TextSensor heatingWarn to display warning messages from the iBoost. At present this only detects the low battery warning from the sensor, but could probably be expanded to include the other messages listed in the manual.